refactor(cache): gather scoped-cache logic into api/src/scoped-cache.ts
Pure structural move (function bodies verbatim) so all scoped-cache logic lives in one module
instead of being split across cache.ts and the giant items.ts — also shrinks the fork's diff
against upstream items.ts and gives pinnedScopeTagsFromCases (branch A) a natural home.

- new api/src/scoped-cache.ts: scopedCachePurgeEnabled, scopedCacheTagKey, tagScopedCacheKeys,
  purgeCache (from cache.ts) + scopedCacheTagsFromRows, pinnedScopeTagsFromFilter (from items.ts)
- cache.ts keeps the general/schema cache; items.ts imports the derivation + purge fns
- repoint importers (respond, tests); split the vi.mock targets so stubs hit scoped-cache.js
  (which ItemsService now imports from), via importOriginal where the real derivation fns must survive

103 unit tests green; no circular imports; gate + directus Lint clean.
