+++ api/src/services/items.ts
@@ -10,7 +10,6 @@ import type {
 	Alterations,
 	Item as AnyItem,
 	EventContext,
-	Filter,
 	MutationTracker,
 	MutationOptions,
 	PrimaryKey,
@@ -24,7 +23,13 @@ import { UserIntegrityCheckFlag } from '@directus/types';
 import type Keyv from 'keyv';
 import type { Knex } from 'knex';
 import { assign, clone, cloneDeep, isPlainObject, omit, pick, without } from 'lodash-es';
-import { getCache, purgeCache, scopedCachePurgeEnabled } from '../cache.js';
+import { getCache } from '../cache.js';
+import {
+	pinnedScopeTagsFromFilter,
+	purgeCache,
+	scopedCacheTagsFromRows,
+	scopedCachePurgeEnabled,
+} from '../scoped-cache.js';
 import { translateDatabaseError } from '../database/errors/translate.js';
 import { getAstFromQuery } from '../database/get-ast-from-query/get-ast-from-query.js';
 import { getHelpers } from '../database/helpers/index.js';
@@ -67,125 +72,6 @@ async function emitActionEvents(actionEvents: ActionEventParams[], opts: Mutatio
 		// an un-awaited rejection (e.g. from a bypassEmitAction handler) doesn't go unhandled.
 		emitting.catch(() => {});
 	}
-}
-
-/**
- * Build scoped cache tags from the distinct scope values present across `rows` — the
- * purge side. With `requireAll` a row missing any scoped cache field makes the values
- * unresolvable (returns `null`), so the caller falls back to a coarse purge rather than
- * leaving a slice stale (a create whose payload omitted the field). Without it, missing
- * fields are skipped — for update payloads where an absent field just means "unchanged"
- * and the pre-update capture covers the old value.
- */
-export function scopedCacheTagsFromRows(
-	collection: string,
-	fields: string[],
-	rows: Record<string, any>[],
-	requireAll: boolean,
-): ScopedCacheTag[] | null {
-	const tags: ScopedCacheTag[] = [];
-
-	for (const field of fields) {
-		const seen = new Set<unknown>();
-
-		for (const row of rows) {
-			if (!(field in row)) {
-				if (requireAll) {
-					return null;
-				}
-
-				continue;
-			}
-
-			const value = row[field];
-
-			if (seen.has(value)) {
-				continue;
-			}
-
-			seen.add(value);
-			tags.push({ collection, field, value });
-		}
-	}
-
-	return tags;
-}
-
-/**
- * Scope a read's root cache tags off the query filter — the read side. A read is
- * soundly scoped to a value slice only when the filter *bounds* it to that value: a
- * future insert with a new scope value must be excluded by the same filter, or the
- * read would silently miss it (a write to the new value purges only its own slice,
- * never this read). So scope tags come from `_eq`/`_in` constraints on a scoped cache
- * field, reached through the root or an `_and` (an `_or` branch doesn't bound — a row
- * matching the other branch still belongs). No pinned field → returns `[]`, and the
- * caller falls back to the bare collection tag so every write to the collection
- * invalidates the read.
- */
-export function pinnedScopeTagsFromFilter(
-	collection: string,
-	fields: string[],
-	filter: Filter | null | undefined,
-): ScopedCacheTag[] {
-	if (!filter || fields.length === 0) {
-		return [];
-	}
-
-	const fieldSet = new Set(fields);
-	const pinned = new Map<string, Set<unknown>>();
-
-	function pin(field: string, values: unknown[]): void {
-		const seen = pinned.get(field) ?? new Set<unknown>();
-
-		for (const value of values) {
-			seen.add(value);
-		}
-
-		pinned.set(field, seen);
-	}
-
-	function walk(node: Filter): void {
-		for (const [key, value] of Object.entries(node)) {
-			if (key === '_and' && Array.isArray(value)) {
-				for (const sub of value) {
-					walk(sub as Filter);
-				}
-
-				continue;
-			}
-
-			// `_or` doesn't bound the read; nothing under it can pin a scope.
-			if (
-				key === '_or' ||
-				!fieldSet.has(key) ||
-				value === null ||
-				typeof value !== 'object'
-			) {
-				continue;
-			}
-
-			const ops = value as Record<string, unknown>;
-
-			if ('_eq' in ops) {
-				pin(key, [ops['_eq']]);
-			}
-			else if ('_in' in ops && Array.isArray(ops['_in'])) {
-				pin(key, ops['_in']);
-			}
-		}
-	}
-
-	walk(filter);
-
-	const tags: ScopedCacheTag[] = [];
-
-	for (const [field, values] of pinned) {
-		for (const value of values) {
-			tags.push({ collection, field, value });
-		}
-	}
-
-	return tags;
 }
 
 export class ItemsService<Item extends AnyItem = AnyItem, Collection extends string = string>
