+++ api/src/services/scoped-cache-purge.test.ts
@@ -36,6 +36,12 @@ const purgeCache = vi.fn();
 vi.mock('../cache.js', () => {
 	return {
 		getCache: () => ({ cache: { clear: vi.fn(), delete: vi.fn() } }),
+	};
+});
+
+vi.mock('../scoped-cache.js', async (importOriginal) => {
+	return {
+		...(await importOriginal<typeof import('../scoped-cache.js')>()),
 		purgeCache,
 		scopedCachePurgeEnabled: () => true,
 	};
